+++ tests/test_filesystem.py
@@ -1,5 +1,4 @@
 import mock
-import pytest
 
 
 def test_filesystem_directory__file_not_found():
@@ -34,3 +33,32 @@ def test_filesystem_directory__returns_fileobject():
     for filename in expected_filepaths:
         with fs.open(filename) as f:
             assert hasattr(f, 'read')
+
+
+def test_filesystem_zip__file_not_found():
+    from pycobertura.filesystem import ZipFileSystem
+
+    fs = ZipFileSystem("tests/dummy/dummy.zip")
+
+    dummy_source_file = 'foo/non-existent-file.py'
+    try:
+        with fs.open(dummy_source_file) as f:
+            pass
+    except ZipFileSystem.FileNotFound as fnf:
+        assert fnf.path == dummy_source_file
+
+
+def test_filesystem_zip__returns_fileobject():
+    from pycobertura.filesystem import ZipFileSystem
+
+    fs = ZipFileSystem("tests/dummy/dummy.zip")
+
+    source_files_in_zip = [
+        'dummy/dummy.py',
+        'dummy/dummy2.py',
+    ]
+
+    for source_file in source_files_in_zip:
+        with fs.open(source_file) as f:
+            assert hasattr(f, 'read')
+    #pycobertura show dummy.original.xml -f html -s dummy/dummy.zip